+++ src/common/b2_settings.cpp
@@ -26,11 +26,12 @@
 #include <cstdio>
 #include <cstdarg>
 #include <cstdlib>
+#include <cstdint>
 
 b2Version b2_version = {2, 4, 1};
 
 // Memory allocators. Modify these to use your own allocator.
-void* b2Alloc_Default(std::int32_t size)
+void* b2Alloc_Default(std::size_t size)
 {
 	return malloc(size);
 }
